backend/pyproject.toml
[project]
name = "nlj-platform-backend"
version = "0.1.0"
description = "FastAPI backend for NLJ Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core FastAPI dependencies
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    # Database
    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
    "asyncpg",
    # Data validation and serialization
    "pydantic",
    "pydantic-settings",
    "email-validator",
    # Authentication
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    # HTTP client for external integrations
    "httpx",
    "aiohttp",
    # Excel file processing (for Trivie imports)
    "openpyxl",
    # Environment management
    "python-dotenv",
    "greenlet>=3.2.3",
    # Claude API integration
    "anthropic",
    # Document conversion
    "python-docx",
    "pypdf",
    "python-magic",
    # File handling
    "aiofiles",
    # Kafka integration
    "aiokafka",
    "kafka-python",
    # Elasticsearch for analytics
    "elasticsearch[async]",
    "bcrypt>=4.3.0",
    # AWS SDK for S3, SES, RDS integration
    "boto3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-postgresql",
    "black",
    "isort",
    "mypy",
    "ruff",
    "autoflake",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
remove-duplicate-keys = true
in-place = true

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "isort>=6.0.1",
    "ruff>=0.12.5",
]
